Extract UiSyncEvent and make available publicly
