+++ ui/appui-react/src/appui-react/syncui/InternalSyncUiEventDispatcher.ts
@@ -6,19 +6,7 @@
  * @module SyncUi
  */
 
-import { BeUiEvent } from "@itwin/core-bentley";
-
-/** UiSync Event arguments. Contains a set of lower case event Ids.
- * @public
- */
-export interface UiSyncEventArgs {
-  eventIds: Set<string>;
-}
-
-/** UiSync Event class.
- * @public
- */
-export class UiSyncEvent extends BeUiEvent<UiSyncEventArgs> { }
+import { UiSyncEvent } from "./UiSyncEvent";
 
 /** This class is used to send eventIds to interested UI components so the component can determine if it needs
  * to refresh its display by calling setState on itself.
