+++ ui/appui-react/src/appui-react/uistate/useUiStateStorage.tsx
@@ -14,7 +14,7 @@ import {
   SyncUiEventDispatcher,
   SyncUiEventId,
 } from "../syncui/SyncUiEventDispatcher";
-import type { UiSyncEventArgs } from "../syncui/InternalSyncUiEventDispatcher";
+import type { UiSyncEventArgs } from "../syncui/UiSyncEvent";
 
 /** @public */
 export function useUiStateStorageHandler(): UiStateStorage {
